requirements.txt
PySide6>=6.5
qtawesome>=1.3
paramiko
pysmb
pyyaml
py7zr
rarfile
pillow
pillow-heif
